Pull up refactoring: Convolve, CustomerConvolveGUI, FilterWithGUI
