+++ src/main/java/pixelitor/colors/Colors.java
@@ -133,34 +133,6 @@ public class Colors {
     }
 
     /**
-     * A linear interpolation for hue values,
-     * taking their circular nature into account
-     */
-    public static float lerpHue(float mixFactor, float hue1, float hue2) {
-        float diff = hue1 - hue2;
-        if (diff < 0.5f && diff > -0.5f) {
-            return ImageMath.lerp(mixFactor, hue1, hue2);
-        } else if (diff >= 0.5f) { // hue1 is big, hue2 is small
-            hue2 += 1.0f;
-            float mix = ImageMath.lerp(mixFactor, hue1, hue2);
-            if (mix > 1.0f) {
-                mix -= 1.0f;
-            }
-            return mix;
-        } else if (diff <= 0.5f) { // hue2 is big, hue1 is small
-            hue1 += 1.0f;
-            float mix = ImageMath.lerp(mixFactor, hue1, hue2);
-            if (mix > 1.0f) {
-                mix -= 1.0f;
-            }
-            return mix;
-        } else {
-            throw new IllegalStateException(
-                format("hue1 = %.2f, hue2 = %.2f, mixFactor = %.2f", hue1, hue2, mixFactor));
-        }
-    }
-
-    /**
      * Just like the above lerpHue method, but taking the average
      * that is on the opposite side of the circle. Not very useful.
      */
